Created a separate project, cadcutils, for the utils library
