+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "astropy_helpers"]
-	path = astropy_helpers
+[submodule "cadcutils/astropy_helpers"]
+	path = cadcutils/astropy_helpers
 	url = https://github.com/astropy/astropy-helpers.git
