+++ cadcutils/MANIFEST.in
