+++ cadcutils/README.md
@@ -1,5 +1,5 @@
-# cadctools
-Client tools for using CADC services
+# cadcutil
+General CADC Python library
 
 Note: The astropy package must be installed before invoking the test command.
 Starting version 1.3.0 of astropy, the test target will support direct
