+++ cadcutils/cadcutils/net/tests/setup_package.py
