+++ cadcutils/cadcutils/tests/coveragerc
