+++ cadcutils/cadcutils/util/tests/coveragerc
