+++ cadcutils/cadcutils/util/tests/setup_package.py
