+++ cadcutils/cadcutils/util/utils.py
