+++ cadcutils/docs/_templates/autosummary/base.rst
