+++ cadcutils/docs/_templates/autosummary/class.rst
