+++ cadcutils/docs/_templates/autosummary/module.rst
