+++ cadcutils/setup.cfg
@@ -21,9 +21,9 @@ doctest_plus = enabled
 auto_use = True
 
 [metadata]
-package_name = cadctools
-description = Core CADC tools 
-long_description = Core tools for accessing CADC Web services
+package_name = cadcutils
+description = Generic CADC Python libary of utilities 
+long_description = Library of utilities. It includes generic utilities for logging and command line parsing, networking utilities, etc.
 author =  Canadian Astronomy Data Centre
 author_email = cadc@nrc-cnrc.gc.ca
 license = AGPLv3
@@ -35,4 +35,4 @@ version = 1.0
 #TODO tests_require = ['astropy'] # this will work in a future version of astropy_helpers (1.3.0)
 
 [entry_points]
-cadc-get-cert = cadctools.net.auth:get_cert_main
+cadc-get-cert = cadcutils.net.auth:get_cert_main
